serviceChecks: refactor to use Checker interface
